Fix test cases
Previous test cases no longer valid.
Updated with new test cases.
