+++ src/test/java/seedu/tasks/DeadlineTest.java
@@ -14,7 +14,7 @@ public class DeadlineTest {
     @Test
     public void test_deadline_DateTimeParseException_thrown() {
         try {
-            new Deadline("go home", "2023/09/02 1000");
+            new Deadline("go home", "29-03-2001");
             fail("Expected InvalidDateException to be thrown");
         } catch (InvalidInputException | InvalidDateException e) {
         }
@@ -22,10 +22,10 @@ public class DeadlineTest {
 
     @Test
     public void testToString() throws InvalidInputException, InvalidDateException {
-        Deadline d = new Deadline("test code", "29-03-2001 0400");
+        Deadline d = new Deadline("test code", "2001-03-29");
         assertEquals("[D][ ] test code (by: MARCH 29 2001)", d.toString());
 
-        Deadline d1 = new Deadline("redo code", "29-03-2001 0500");
+        Deadline d1 = new Deadline("redo code", "2001-03-29");
         assertEquals("[D][ ] redo code (by: MARCH 29 2001)", d1.toString());
     }
 }
